requirements.txt
gspread==5.5.0
Jinja2==3.0.1
oauth2client==4.1.3
pandas==1.3.2
pdfkit==1.0.0
Pillow==9.3.0
streamlit==1.13.0
